requirements.txt
PyMuPDF==1.24.9
pdfminer.six==20231228
typing_extensions>=4.9.0
python-dotenv==1.0.1
Flask==3.0.3

# OCR
pytesseract==0.3.13
Pillow==10.4.0

